chat/script: !uids.indexOf(USER_ID) casted to Number
uids.indexOf(USER_ID) returns either 1 or 0 as the index.
using ! inverts the value. if USER_ID is at position 0,
other user's id is at position 1 which is == !0.

but instead of 1, !0 returns true. Array[Boolean] returns
undefined.

Array[Number(Boolean)] casts the value to a Number, which is
a valid index.