chat/script: replaced onValue with onChildAdded
This function executes seperately for each child element
i.e. a message on session start, which allows the callback to
seperately load the messages one by one.

Later during a session, for every child added, only that specific
child i.e. a message is fetched from the database, instead of the
entire set of chats, which improves performance.