chat/script: added onChildRemoved function
Since the onValue funxtion got removed in the last commit by
onChildAdded, only additions will be fetched.

This means unsent messages can't be tracked by onChildAdded.

This is where onChildRemoved comes into play. This function
fetches the message which got removed, and usong this data,
the message is located bybits pushkey in the UI, and that HTML
node is deleted.