chat/script: the temporary placeholder nodes added during update contain html id place-holder
A placeholder node was added during update to make up for the latency
of data being sent to the database, and the data being fetched from
and loaded into the UI.

The id to this was unnecessary when the whole db would
be overwritten by onValue.

But now messages are seperately appended, which means the real message
will get appended after the placeholder, making duplications of
each message.

So, the placeholder is removed by loadMessagesToUI function
before appending a real message. This is possible because at any given
moment there can be only one placeholder node.

To remove the place holder, it is reffered to by the id 'place-holder'.